Store span start time as an int32 for millisecond precision
We need v3 of the Athena query engine to get query reuse but it has
started enforcing timestamp precision. Previously, it was stored as an
int64 but in parquet (or at least the parquet library we use) that is
microsecond precision. This is why no trace IDs are being returned
because we try to filter by `start_time` which isn't serialisable.